organising/adding mvvm on add a blog view
